Catchup on nightly clippy lints (#768)
These were all fixed up in #759, but the build noise is getting a bit
much on independent PRs so I'm pulling them out and making sure things
still compile via this PR. Everything here is a result of following
clippy's commands (or cleanup when doing so).

Broadly, these are all of the form `format!("{}", s)` ->
`format!("{s}")`.

We have not changed the OPTE API in any way, so I will ignore CI
complaining about `API_VERSION`.