ICMP code/type firewall filters (#759)
This PR changes how Oxide VPC firewall protocol matches are specified,
such that we can now filter on (ranges of) ICMP types/codes. This allows
Nexus and instances to allow limited inbound ICMP traffic where needed
to enable correct operation of Path MTU discovery.

One of the main pieces which needed to be accounted for here was the
shape of an `InnerFlowId` when handling ICMP traffic. Before, both the
`source_port` and `destination_port` would be used to pull double-duty
as storage for ICMP echo IDs. Having this information in the flowkey in
some form is necessary to enable outbound `ping` in a SNAT context. This
works great in a world where we treat all ICMP homogenously. However,
this would enable situations where we create an pathway for *all*
subsequent ICMP flavours from the remote host.

* `InnerFlowId` now contains a `[u16; 2]` which is reinterpreted based
on the protocol. For TCP/UDP, these are just the ports as before. In
ICMP, we have the type/code/echo_id.
* `mirror`ed flow IDs now behave a bit more carefully in this regard.
ICMP(v6) echo request/reply change the expected `type` value at this
step and keep the same `echo_id` in place.

While I've been in this code, I've taken the opportunity to summarise
port matches for more typical firewall rules, since that was an
`O(n_values)` operation in the slowpath.

Closes #730.
Closes #468.