Use `omicron1` brand in CI (#782)
Very quick PR to enable use of the `omicron1` brand in the `kbench
local` benchmark, and the xde simnet connectivity test. This shaves a
good 6 min off `xde-test` and `bench` in CI -- since sparse zones take
so long to setup that it's worth spending ~1min to build the baseline
for omicron1 zones.

Basically, this zone type requires that we put a VNIC ontop of each
simnet endpoint and then use *that* as the underlay. This is also, as it
turns out, a requirement to use packet siphons in these test cases.