Don't apply underlay boost to poor MSS choices
Clients expect us to have the correct MSS for the link signalled in the
output packet when we perform pseudo-GRO. Ordinarily, the receive side
is making the assumption when it sees an oversize packet that this would
be 1500B, however there are cases where the guest has chosen a reduced
MSS for tunnel-related reasons (either correctly or incorrectly). This
information is effectively lost, and if the guest attempts to rely on
this then things won't work as intended.

This PR breaks such packets up into the frame sizes demanded by the
guest if the MSS is lower than that expected for `1500 - hdrs_len`.