Add a mechanism for explicit validation of push specs
Push specifiers for encap and IP work on a slightly simplified view of
the world. 'Til now, we've had no need to explicitly validate what a
`NetworkImpl` gives us -- these elements are fixed-size, and we fill in
the lengths without any fuss.

Now that implementors can ask for Options and Extension headers as part
of this spec, we need to actually validate that these things can go on
the wire (e.g., no per-option or per-extension payloads which go beyond
the maximum length of its respective container).

This is going to be mecessary and useful for the Geneve TLVs we want to
support as part of multicast.