Wherein we Teach OPTE about The Origin of Geneve Options
Overall, OPTE now carries generic geneve options as part of both its
owned and borrowed representations for encap metadata. One of the things
which bothered me about how we've been doing this (both pre- and
post-ingot) is that our understanding of some options is baked into the
metadata format itself, when in actual fact this needs to be
interpreted by the `NetworkImpl`, *not* the engine. This will also be
more useful once update is on the scene, and nodes may well be
communicating over different OPTE engine versions (thus, different
support for whatever our option codes are).

Nodes on `glasgow` remain mutually intelligible over TCP using this new
extension, and making space to carry the MSS over geneve.

I need to pull up the Geneve option parsing/interpretation code into the
`oxide-vpc` crate to fix the external-tag handling, and give us
something consistent for both `NetworkImpl` and XDE code. One of the
tests will be broken until this is the case.