Send MSS in-band when performing pseudo-GRO (#805)
For reasons which aren't entirely clear yet, Linux guests using `podman`
+ DNAT with a reduced MTU for the contained network will reject GRO'd
packets iff. the `gso_size` is larger than their MTU. When performing
GRO normally, a NIC would extract this parameter from the largest TCP
segment seen when assembling a larger frame. For rack-internal traffic,
we boost the MSS to that of a 9k MTU, wherein we obviously lose
knowledge of the `gso_size`/MSS. We currently infer `gso_size` using a
1500B MTU and the size of the carried headers -- this works assuming
that neither guest reduces this internally. So, in the motivating case,
any LSO'd packets going between sleds end up getting dropped at the
receiving end.

To propagate this information correctly, we now allocate a Geneve Option
header (option type `0x02` in our typical class) as part of the
encapsulation of any rack-internal TCP traffic. When we send such
packets to the underlay, we check to see whether a) the packet expects
LSO and b) whether it has a nonzero MSS. If so, we write the original
MSS value into the destination option. The recipient looks for this
marker, and when seen compares it against the payload length to
correctly mark frames before they go to the guest.

---

To do so, this PR standardises the machinery for expressing Geneve and
IPv6 extensions within layer push actions, and drastically simplifies
the code needed to walk Geneve options using a given parser
implementation. Both of these option-related systems have been written
here because the initial version of the fix relied instead upon an
IPv6EH to avoid the [requisite switch
changes](https://github.com/oxidecomputer/dendrite/pull/82).

Supersedes #801. That change negates the pseudo-GRO performance benefits
on low-MTU traffic by splitting packets *exactly as the guest demanded*,
which has severe performance impact. This PR has another benefit in that
it should also correctly label packets if the OPTE/viona MTU are
increased in future.

Fixes https://github.com/oxidecomputer/customer-support/issues/480.