Don't allocate in `BTreeMap` when only one postbox is needed (#819)
We keep track of `MsgBlkChain`s in a `BTreeMap` today. Sadly this isn't
a type which we can preallocate a desired capacity in and store per-CPU.
As a result, for flamegraphs in the Rx pathway we can see a moderately
chunky call into `kmem_alloc` within `Postbox::post` for chains which
have one recipient.

This PR special-cases the single-recipient case so that we hold onto one
chain and key on the heap until needed.