`DevMap` should not expose cloneable references (#882)
One of the purposes of `DevMap` is to ensure that the datapath does not
inappropriately clone out and extend the lifetime of any ports. As it
turns out, the code using `DevMap` is doing the right thing, but its API
surface is overly permissive and should be making an `Arc::clone` or
similar impossible.